Move abstract CreationState to "_" directory.
